pp365-shared@1.7.3-894.1: Updated ListLogger
