+++ SharePointFramework/@Shared/package.json
@@ -1,6 +1,6 @@
 {
   "name": "pp365-shared",
-  "version": "1.7.3-892.2",
+  "version": "1.7.3-894.1",
   "repository": {
     "type": "git",
     "url": "https://github.com/Puzzlepart/prosjektportalen365.git",
