--- SharePointFramework/@Shared/src/logging/IListLoggerEntry.ts
@@ -1,10 +1,0 @@
-import { ListLoggerEntryLevel } from './ListLoggerEntryLevel'
-
-export class IListLoggerEntry {
-  webUrl?: string
-  scope?: string
-  functionName?: string
-  message: string
-  level?: ListLoggerEntryLevel
-  component?: string
-}
