--- SharePointFramework/@Shared/src/logging/IListLoggerMemberMap.ts
@@ -1,9 +1,0 @@
-export class IListLoggerMemberMap {
-  webUrl?: string
-  scope?: string
-  functionName?: string
-  message?: string
-  level?: string
-  component?: string
-  context?: string
-}
