+++ SharePointFramework/@Shared/src/logging/ListLogger.ts
@@ -1,18 +1,11 @@
 /* eslint-disable max-classes-per-file */
 import { ItemAddResult, List } from '@pnp/sp'
-import { IListLoggerEntry } from './IListLoggerEntry'
-import { IListLoggerMemberMap } from './IListLoggerMemberMap'
-import { ListLoggerEntryLevel } from './ListLoggerEntryLevel'
-
-export const defaultListLoggerMemberMap: Record<string, string> = {
-  webUrl: 'GtLogWebUrl',
-  scope: 'GtLogScope',
-  functionName: 'GtLogFunctionName',
-  message: 'GtLogMessage',
-  level: 'GtLogLevel',
-  component: 'GtLogComponentName',
-  context: 'GtLogContext'
-}
+import {
+  defaultListLoggerMemberMap,
+  IListLoggerEntry,
+  IListLoggerMemberMap,
+  ListLoggerEntryLevel
+} from './types'
 
 class ListLogger {
   public list: any
@@ -21,7 +14,7 @@ class ListLogger {
   public scope: string = ''
 
   /**
-   * Initialize ListLogger
+   * Initialize `ListLogger`
    *
    * @param list List
    * @param webUrl Web URL
@@ -78,11 +71,11 @@ class ListLogger {
   }
 
   /**
-   * Get sp item for entry
+   * Get SP item properties for entry
    *
    * @param entry Entry
    */
-  private _getSpItem(entry: IListLoggerEntry) {
+  private _getSpItem(entry: IListLoggerEntry): Record<string, any> {
     let item: Record<string, any> = {}
     item = Object.keys(this.memberMap).reduce((_item, key) => {
       const fieldName = this.memberMap[key]
@@ -95,8 +88,6 @@ class ListLogger {
     if (this.scope && this.memberMap.scope) {
       item[this.memberMap.scope] = this.scope
     }
-    // eslint-disable-next-line no-console
-    console.log(this, item)
     return item
   }
