--- SharePointFramework/@Shared/src/logging/ListLoggerEntryLevel.ts
@@ -1,1 +1,0 @@
-export type ListLoggerEntryLevel = 'Info' | 'Warning' | 'Error'
